Fixed major bug and finished project
This project checks if a given board ( a dictionary) is valid based on certain conditions.

Conditions are that the key (coordinate) must be a validate coordinate on chessboard. And the pieces cant exceed 32, and that black cant have 16+ pieces and white cant have 16+ pieces and that neither side can have more than 8 pawns